feat: bitwise_and operation added

+++ src/operators/tensor/implementations/tensor_i32.cairo
@@ -245,7 +245,7 @@ impl I32Tensor of TensorTrait<i32> {
         math::where::where(self, x, y)
     }
 
-    fn bitwise_and(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<usize> {
+    fn bitwise_and(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<i32> {
         math::bitwise_and::bitwise_and(self, other)
     }
 }
